Correction of Julia wrapper and Julia test

+++ include/emdee.f03
@@ -192,9 +192,9 @@ interface
   end function EmDee_dihedral_harmonic
 
   subroutine EmDee_Rotational_Energies( md, Kr ) bind(C,name="EmDee_Rotational_Energies")
-    import :: tEmDee, rb
-    type(tEmDee), value   :: md
-    real(rb), intent(out) :: Kr(3) 
+    import :: tEmDee, c_double
+    type(tEmDee),   value       :: md
+    real(c_double), intent(out) :: Kr(3) 
   end subroutine EmDee_Rotational_Energies
 
 end interface
